Deep_Learning/ArtForgeAI/requirements.txt
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
torch>=1.9.0
matplotlib>=3.4.0
pygame>=2.0.0
gym>=0.21.0
stable-baselines3>=1.5.0
pytest>=6.2.0
pytest-cov>=2.12.0
